Added, extended and completed many Javadoc comments, especially parameters and return values.
Added error handling for NumberFormatExceptions that could occur when an 
invalid input was passed for the port number, causing the application to
crash with fatal error.

Corrected processing for help commands. Help for logLevel and quit
commands was missing, help for disconnect command was duplicated.

Added some more log entries.
I'll hand in the project at 22:00. If you want more time please drop me
an email!